fix(api): nit fixes
refs. metriport/metriport-internal#2040
